Ввёл корректировки
